# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestor Recursos API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80/api/v1"
        }
    ],
    "paths": {
...
        "/youtube-subscription?searchTags=[]&searchTitle=''&page=1&perPage=5": {
            "get": {
                "tags": [
                    "YoutubeSubscription"
                ],
                "summary": "List Youtube Subscription",
                "description": "List Youtube Subscription",
                "operationId": "getYoutubeSubscription",
                "parameters": [
                    {
                        "name": "searchTags",
                        "in": "query",
                        "description": "Array of Tag's id",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "number"
                            }
                        }
                    },
                    {
                        "name": "searchTitle",
                        "in": "query",
                        "description": "Name to filter recourses",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of page of pagination",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Amount of register for page",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "identificador": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "youtubeId": {
                                                        "type": "string",
                                                        "example": "SkZG2_P243X9PONV3QdZukxHB-ASrwsvbB5CvjuovAs"
                                                    },
                                                    "usuarioId": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "canalId": {
                                                        "type": "string",
                                                        "example": "UCoQPEruJuFCppnu2fBtyEbA"
                                                    },
                                                    "titulo": {
                                                        "type": "string",
                                                        "example": "El Loco Pildorita"
                                                    },
                                                    "fechaSubscripcion": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2023-09-28"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Hola soy el LOCO PILDORITA, te envito a que te suscribas a mi canal, activa la campanita, dale me gusta, comparte. NO TE ARREPENTIRAS. TU AMIGO EL LOCO PILDORITA\\n#ellocopildorita #comicosambulantes #tendencia"
                                                    },
                                                    "fotoDefault": {
                                                        "type": "string",
                                                        "example": "https://yt3.ggpht.com/x90MXeHCLyZeGOvdVDqOjegCSpKbvVf-ZsVH_jHz16HowpXuENro3KrJSrfaicJiH9UZN315sw=s88-c-k-c0x00ffffff-no-rj"
                                                    },
                                                    "fotoMedium": {
                                                        "type": "string",
                                                        "example": "https://yt3.ggpht.com/x90MXeHCLyZeGOvdVDqOjegCSpKbvVf-ZsVH_jHz16HowpXuENro3KrJSrfaicJiH9UZN315sw=s240-c-k-c0x00ffffff-no-rj"
                                                    },
                                                    "fotoHigh": {
                                                        "type": "string",
                                                        "example": "https://yt3.ggpht.com/x90MXeHCLyZeGOvdVDqOjegCSpKbvVf-ZsVH_jHz16HowpXuENro3KrJSrfaicJiH9UZN315sw=s800-c-k-c0x00ffffff-no-rj"
                                                    },
                                                    "tags": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "identificador": {
                                                                    "type": "integer",
                                                                    "example": 2
                                                                },
                                                                "nombre": {
                                                                    "type": "string",
                                                                    "example": "PROGRAMACIÓN"
                                                                },
                                                                "estilos": {
                                                                    "type": "string",
                                                                    "example": "bg-gray-900 text-white"
                                                                },
                                                                "total": {
                                                                    "type": "integer",
                                                                    "example": 0
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag"
                                                },
                                                "currentPage": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "perPage": {
                                                    "type": "number",
                                                    "example": 5
                                                },
                                                "totalPages": {
                                                    "type": "number",
                                                    "example": 3
                                                },
                                                "from": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "number",
                                                    "example": 5
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "links": {
                                            "properties": {
                                                "self": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=1"
                                                },
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=4"
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=2"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/youtube-subscription/checkstatus": {
            "get": {
                "tags": [
                    "YoutubeSubscription"
                ],
                "summary": "Get Status of proccess to import YoutubeSubscription",
                "description": "Get Status of proccess to import YoutubeSubscription",
                "operationId": "getYoutubeSubscriptionChekcStatus",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "procesando"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/youtube-subscription": {
            "post": {
                "tags": [
                    "YoutubeSubscription"
                ],
                "summary": "Import Youtube Subscription from Google Login OAuth",
                "description": "Import Youtube Subscription from Google Login OAuth",
                "operationId": "postYoutubeSubscription",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "access_token",
                                    "order"
                                ],
                                "properties": {
                                    "author": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "editorial": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "access_token": "ASK29ZXC_aSD1741%5&4ECS",
                                    "order": "relevance"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Se insertaron los registros"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}